backend/requirements.txt
fastapi == 0.134.0 
uvicorn == 0.41.0 
pinecone == 8.1.0 
onnxruntime-gpu == 1.23.2
numpy == 2.2.6  
openai == 2.24.0
pyyaml == 6.0.3 
pydantic-settings == 2.13.1 
transformers == 5.2.0
pinecone-text == 0.11.0
huggingface-hub == 1.5.0  